+++ src/drv/analog/adc_task_function.cpp
@@ -31,6 +31,10 @@ namespace drv::analog {
 
 namespace {
 
+std::array<uint8_t, conversion_frame_size> conversion_frame;
+VoltagesQueue::value_type voltages;
+CurrentsQueue::value_type currents;
+
 /// Convert the NVS setting "cur_sc_time" to a counter value
 ///
 /// For convenience, the \ref mem::nvs::Settings::getCurrentShortCircuitTime()
@@ -44,6 +48,39 @@ auto get_short_circuit_count() {
 }
 
 } // namespace
+
+/// Handles suspend/resume logic
+///
+/// Since the ADC task holds a mutex on the [continuous mode
+/// driver](https://docs.espressif.com/projects/esp-idf/en/\idf_ver/esp32s3/api-reference/peripherals/adc_continuous.html#_CPPv427adc_continuous_handle_cfg_t),
+/// the only way to stop and start the ADC is to send a task notification. This
+/// is done using adc_task_notify_suspend() or adc_task_notify_resume(). The
+/// conversion frame then starts again from the beginning and this method can be
+/// used to synchronize the ADC.
+void handle_suspend_resume_on_notify() {
+  uint32_t notified_value{};
+  xTaskNotifyWaitIndexed(default_notify_index, 0u, 0u, &notified_value, 0u);
+  if (notified_value) {
+    ESP_ERROR_CHECK(adc_continuous_stop(adc1_handle));
+    ESP_ERROR_CHECK(adc_continuous_flush_pool(adc1_handle));
+    xTaskNotifyWaitIndexed(
+      default_notify_index, 0u, ULONG_MAX, &notified_value, portMAX_DELAY);
+    assert(!notified_value);
+    ESP_ERROR_CHECK(adc_continuous_start(adc1_handle));
+  }
+}
+
+/// Suspend ADC task with task notification
+void adc_task_notify_suspend() {
+  xTaskNotifyIndexed(
+    adc_task.handle, default_notify_index, 1u, eSetValueWithOverwrite);
+}
+
+/// Resume ADC task with task notification
+void adc_task_notify_resume() {
+  xTaskNotifyIndexed(
+    adc_task.handle, default_notify_index, 0u, eSetValueWithOverwrite);
+}
 
 /// ADC task function
 ///
@@ -59,7 +96,6 @@ auto get_short_circuit_count() {
 /// "short circuit" and a \ref page_mw_roco track short circuit message is
 /// broadcast.
 [[noreturn]] void adc_task_function(void*) {
-  std::array<uint8_t, conversion_frame_size> conversion_frame{};
   auto short_circuit_count{get_short_circuit_count()};
 
   // Start and stop must be called from the same task because the handle uses a
@@ -80,9 +116,7 @@ auto get_short_circuit_count() {
       continue;
     }
 
-    // The following conversion and copy takes 20.4µs
-    VoltagesQueue::value_type voltages;
-    CurrentsQueue::value_type currents;
+    // The following conversion and copy takes 174us
     auto voltages_it{begin(voltages)};
     auto currents_it{begin(currents)};
     for (auto i{0uz}; i < bytes_received; i += SOC_ADC_DIGI_RESULT_BYTES) {
@@ -112,11 +146,10 @@ auto get_short_circuit_count() {
     // Clear count if no short circuit
     else
       short_circuit_count = get_short_circuit_count();
-  }
 
-  // Start and stop must be called from the same task because the handle uses a
-  // FreeRTOS mutex for internal locking
-  ESP_ERROR_CHECK(adc_continuous_stop(adc1_handle));
+    // Handle suspend/resume on notify
+    handle_suspend_resume_on_notify();
+  }
 }
 
 } // namespace drv::analog